works. splitted interface and configs. added new type for schedule input
