+++ frame/composable-traits/src/undercollateralized_loans/configs.rs
@@ -1,4 +1,4 @@
-use crate::defi::{CurrencyPair, DeFiEngine};
+use crate::defi::CurrencyPair;
 use composable_support::math::safe::SafeAdd;
 use frame_support::pallet_prelude::*;
 use sp_runtime::{traits::Zero, ArithmeticError, Perquintill};
@@ -116,7 +116,7 @@ where
 	interest: Percent,
 	/// How often borrowers have to pay interest.
 	payment_frequency: BlockNumber,
-	/// Activated loan lifetime in the terms of block numbers.
+    /// Activated loan lifetime in the terms of block numbers.
 	maturity: BlockNumber,
 	/// Payment strategie which should be applyed.
 	/// For instance borrower have to pay principal when loan is mature (one strategy),
@@ -321,8 +321,8 @@ impl<AccountId, AssetId: Copy, BlockNumber, LiquidationStrategyId>
 	}
 }
 
-#[derive(Encode, Decode, Default, TypeInfo, RuntimeDebug, Clone, PartialEq)]
-pub struct LoanInput<AccountId, Balance, BlockNumber, Percent, RepaymentStrategy> {
+#[derive(Encode, Decode, TypeInfo, Clone, PartialEq, RuntimeDebug)]
+pub struct LoanInput<AccountId, Balance, BlockNumber, Date, Percent, RepaymentStrategy> {
 	/// Loan belongs to this market.
 	pub market_account_id: AccountId,
 	/// This account id have to be whitelisted.
@@ -335,6 +335,7 @@ pub struct LoanInput<AccountId, Balance, BlockNumber, Percent, RepaymentStrategy
 	pub interest: Percent,
 	/// How often borrowers have to pay interest.
 	pub payment_frequency: BlockNumber,
+    pub payment_schedule: Vec<Payment<Date, Percent>>,
 	/// Loan shoud be paid back after this amount of blocks.
 	pub loan_maturity: BlockNumber,
 	/// Payment strategie which should be applyed.
@@ -343,73 +344,10 @@ pub struct LoanInput<AccountId, Balance, BlockNumber, Percent, RepaymentStrategy
 	pub repayment_strategy: RepaymentStrategy,
 }
 
-pub trait UndercollateralizedLoans: DeFiEngine {
-	type BlockNumber: Clone + Eq + PartialEq;
-	type LiquidationStrategyId: Clone + Eq + PartialEq;
-	type Percent: Clone + Eq + PartialEq;
-	type VaultId: Clone + Eq + PartialEq;
-	type RepaymentStrategy: Clone;
-
-	fn create_market(
-		manager: Self::AccountId,
-		input: MarketInput<
-			Self::AccountId,
-			Self::MayBeAssetId,
-			Self::BlockNumber,
-			Self::LiquidationStrategyId,
-		>,
-		keep_alive: bool,
-	) -> Result<
-		MarketInfo<
-			Self::AccountId,
-			Self::MayBeAssetId,
-			Self::BlockNumber,
-			Self::LiquidationStrategyId,
-			Self::VaultId,
-		>,
-		DispatchError,
-	>;
-
-	fn create_loan(
-		input: LoanInput<
-			Self::AccountId,
-			Self::Balance,
-			Self::BlockNumber,
-			Self::Percent,
-			Self::RepaymentStrategy,
-		>,
-	) -> Result<
-		LoanConfig<
-			Self::AccountId,
-			Self::Balance,
-			Self::BlockNumber,
-			Self::Percent,
-			Self::RepaymentStrategy,
-		>,
-		DispatchError,
-	>;
-
-	fn borrow(
-		borrower_account_id: Self::AccountId,
-		loan_account_id: Self::AccountId,
-		keep_alive: bool,
-	) -> Result<
-		LoanInfo<
-			Self::AccountId,
-			Self::Balance,
-			Self::BlockNumber,
-			Self::Percent,
-			Self::RepaymentStrategy,
-		>,
-		DispatchError,
-	>;
-
-	fn market_account_id<S: Encode>(postfix: S) -> Self::AccountId;
-
-	fn loan_account_id<S: Encode>(postfix: S) -> Self::AccountId;
-
-	fn is_borrower_account_whitelisted(
-		borrower_account_id_ref: &Self::AccountId,
-		market_id_ref: &Self::AccountId,
-	) -> Result<bool, DispatchError>;
+#[derive(Encode, Decode, TypeInfo, Clone, PartialEq, RuntimeDebug)]
+pub struct Payment<Date, Percent> {
+    pub date: Date,
+    pub rate: Percent,
 }
+
+
